Adding matcher to the health check.
The matcher is the matcher for the status code. The current code will
end up with a default of 200 in AWS but this should be specified for any
health check for an ALB.

This is useful for when services have a login page and you are checking
the root or other similar cases.

Updated the documentation and the code.